refactor: mudança na NeuralNetwork de double para float

+++ include/NeuralNetwork/NNet.h
@@ -7,18 +7,18 @@ class NNet {
     public:
         NNet();
 
-        const double&  feedFoward(vec& input);
-        void backPropagation(double target, double learnRate);
+        const float&  feedFoward(vec& input);
+        void backPropagation(float target, float learnRate);
 
     private:
         mat layer1;
         vec layer2;
         vec in;
         vec outL1;
-        double outL2;
+        float outL2;
 };
 
 namespace nnet {
-    double Tanh(double x, bool derivate = false);
-    double normalizeTanh(double value, double min, double max);
+    float Tanh(float x, bool derivate = false);
+    float normalizeTanh(float value, float min, float max);
 }
